+++ apps/web/src/app/test-form/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import MultiStepForm from '@/components/common/multi-step-form'
+import { MultiStepForm } from '@/components/forms'
 
 export default function TestFormPage() {
   const handleFormSubmit = (data: any) => {
