+++ apps/web/src/components/forms/index.ts
@@ -1,2 +1,3 @@
 // Export all form components
+export * from './multi-step'
 export * from './sustainability'
