+++ apps/web/src/components/forms/multi-step/components/step-indicator.tsx
