+++ apps/web/src/components/forms/multi-step/index.ts
@@ -1,5 +1,8 @@
 // Export all schemas and configurations
 
+// Export main component
+export { default as MultiStepForm } from './multi-step-form'
+
 // Export all components
 export * from './components'
 // Export all hooks
